tutorials/netopanalytics/edge-subsystem/Compose-based/mini-batch-converter/requirements.txt
paho-mqtt==1.5.0
msgpack==1.0.0
kafka-python==2.0.1